server/requirements.txt
flask>=3.0
google-generativeai>=0.8
python-dotenv>=1.0
pyserial>=3.5
